added all current server datapacks
Added all of the datapacks currently being used on the server beside the vanillatweaks datapacks, this includes all origin datapacks either custom made by someone on the server or downloaded from somewhere else and put on the server. These datapacks also include the custom function datapacks I added, the foxs_car and invsee datapacks were added. The resourcepack for the foxs_car dtapack should also be on the server, if not please dm me at FoxMaster198#5113